Refactor PSU module: Split strategies, rename files, add deprecation warning
